rename deserialize signature and fix double negative condition
