+++ frontend/docs/frontend/serializing-state-to-url.mdx
@@ -52,7 +52,7 @@ This method expects a configuration object with the following signature:
 const deserializeConfig = {
   props: Array,
   params: Object,
-  initialState: Object,
+  state: Object,
   urlPropHandlers: {
     [propName]: {
       parse: value => parsedValue,
