+++ frontend/scripts/react-components/shared/url-serializer/url-serializer.js
@@ -3,20 +3,7 @@ import { redirect } from 'redux-first-router';
 import omit from 'lodash/omit';
 import UrlSerializer from './url-serializer.component';
 
-export const DONT_SERIALIZE = 'NO_CEREAL_RYAN_GOSLING';
-
-export const deserialize = ({ props, params, initialState = {}, urlPropHandlers = {} }) =>
-  props.reduce((acc, prop) => {
-    const getParsedValue = urlPropHandlers[prop] ? urlPropHandlers[prop].parse : x => x;
-    return {
-      ...acc,
-      [prop]:
-        typeof params[prop] !== 'undefined' ? getParsedValue(params[prop]) : initialState[prop]
-    };
-  }, initialState);
-
 const mapStateToProps = state => ({
-  DONT_SERIALIZE,
   query: state.location.query
 });
