+++ frontend/scripts/react-components/tool/tool-search/tool-search.saga.js
@@ -1,7 +1,7 @@
 import { SET_SELECTED_NODES_BY_SEARCH } from 'react-components/tool/tool.actions';
 import { all, fork, takeLatest, select, put } from 'redux-saga/effects';
-import { getVisibleNodes } from 'react-components/tool/tool.selectors';
 import { expandSankey } from 'react-components/tool-links/tool-links.actions';
+import { getVisibleNodes } from 'react-components/tool-links/tool-links.selectors';
 
 function* checkExpandNode() {
   function* performCheckexpandNode({ payload }) {
